requirements.txt
streamlit>=1.45.0,<2
google-genai>=1.16.0,<3
huggingface_hub>=0.32.0,<1.0
Pillow>=10.0.0
